Enable device list when user access first time
