+++ src/AudienceUi.jsx
@@ -2,8 +2,10 @@ import React, { Component } from 'react';
 import { IndexRoute, Link, Router, Route, browserHistory } from 'react-router';
 import RemoteVideos from './RemoteVideos';
 import LocalVideo from './LocalVideo';
+import getCameraStream from './getCameraStream';
+import connectToSkyWay from './connectToSkyWay';
+import joinRoom from './joinRoom';
 import CONST from './Const';
-import webinar from './webinar';
 import './AudienceUi.css';
 
 class AudienceUi extends Component {
@@ -14,9 +16,28 @@ class AudienceUi extends Component {
             { op: 'replace', path: '/mode', value: CONST.ROLE_AUDIENCE },
             { op: 'replace', path: '/roomName', value: this.props.params.roomName}
         ]);
+
+        function _webinar(myPeerId, width, height, frameRate, isMuted) {
+            const _joinRoom = joinRoom.bind(this);
+            connectToSkyWay.bind(this)(myPeerId, () => {
+                const isFinishedGetCameraStream = !!(this.props.localStream);
+                if (!isFinishedGetCameraStream) {
+                    return;
+                }
+                _joinRoom();
+            });
+            getCameraStream.bind(this)(width, height, frameRate, isMuted, () => {
+                const isFinishedConnectToSkyWay = !!(this.props.peer);
+                if (!isFinishedConnectToSkyWay) {
+                    return;
+                }
+                _joinRoom();
+            });
+        }
+
         if (!this.isWebinarStarted) {
             this.isWebinarStarted = true;
-            webinar.bind(this)(null, CONST.AUDIENCE_CAMERA_WIDTH, CONST.AUDIENCE_CAMERA_HEIGHT, CONST.AUDIENCE_CAMERA_FRAME_RATE, true);
+            _webinar.bind(this)(null, CONST.AUDIENCE_CAMERA_WIDTH, CONST.AUDIENCE_CAMERA_HEIGHT, CONST.AUDIENCE_CAMERA_FRAME_RATE, true);
         }
     }
     render() {
